+++ tests/testthat/test_extractData.R
@@ -54,10 +54,11 @@ test_that("char2fac", {
   df <- data.frame(v1 = factor(c("z", "a", "b"), levels = c("z", "a", "b")),
                    stringsAsFactors = TRUE)
   gads <- import_DF(df)
-
   dat <- extractData(gads, convertLabels = "character")
+
   out <- char2fac(dat, labels = gads$labels, vars = "v1", convertMiss = TRUE)
   expect_true(is.factor(out$v1))
+  expect_false(is.ordered(out$v1))
   expect_equal(as.numeric(out$v1), c(1:3))
 })