+++ NEWS.md
@@ -1,5 +1,6 @@
 # eatGADS 0.20.0.9000
 
+* `extractData2()` for better usability and support of ordered factors
 * `fixEncoding()` now supports imports of `windows-1250` imported files via `import_spss()` with `UTF-8` encoding
 * `import_spss()` now transforms `DATEATIME` variables to character, assigning an appropriate format
 * `cloneVariable()` for duplicating an existing variable under a new name